MIDIIn: now handles poly input like VCV Rack.
also major UI update to handle voice and poly mode config

+++ software/src/applets/hMIDIOut.h
@@ -245,7 +245,7 @@ private:
     MIDILogEntry log[7];
     int log_index;
 
-    void UpdateLog(int message, int data1, int data2) {
+    void UpdateLog(uint8_t message, uint8_t data1, uint8_t data2) {
         log[log_index++] = {message, data1, data2};
         if (log_index == 7) {
             for (int i = 0; i < 6; i++) {
